Move handler function to handlers package
